requirements.txt
ansiblemetrics==0.1
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.1
coverage==4.5.4
docutils==0.15.2
importlib-metadata==0.23
joblib==0.14.0
more-itertools==7.2.0
numpy==1.17.4
pandas==0.25.3
parameterized==0.7.0
pluggy==0.13.0
py==1.8.0
pytest==4.5.0
pytest-cov==2.7.1
python-dateutil==2.8.1
python-interface==1.5.1
pytz==2019.3
PyYAML==5.1
scikit-learn==0.21.2
scipy==1.3.2
six==1.13.0
sklearn==0.0
statistics==1.0.3.5
wcwidth==0.1.7
zipp==0.6.0
